AppCons180 Modificar ingreso de pedido por cliente

+++ app/Portal.Consultoras.Web/Areas/Mobile/Models/MobileAppConfiguracionModel.cs
@@ -13,7 +13,7 @@ namespace Portal.Consultoras.Web.Areas.Mobile.Models
         /// Establece si se mostrara o no el boton atras en las vistas
         /// </summary>
         public bool MostrarBotonAtras { get; set; }
-        public long ClienteID { get; set; }
+        public int ClienteID { get; set; }
 
         public MobileAppConfiguracionModel()
         {
